pip_freeze_requirements.txt
attrs==25.3.0
Automat==25.4.16
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.5
cssselect==1.3.0
cycler==0.12.1
defusedxml==0.7.1
filelock==3.18.0
fonttools==4.58.5
fsspec==2024.6.1
huggingface-hub==0.33.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
lxml==6.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pandas==2.3.1
parsel==1.10.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
queuelib==1.8.0
requests==2.32.4
requests-file==2.1.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
Scrapy==2.13.3
seaborn==0.13.2
service-identity==24.2.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.17
tldextract==5.3.0
torch==2.5.0+cu118
torchaudio==2.5.0+cu118
torchvision==0.20.0+cu118
tqdm==4.67.1
Twisted==25.5.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
w3lib==2.3.1
zope.interface==7.2
